Move binary search to own file
